+++ examples/geojson-animation/src/app.js
@@ -21,32 +21,17 @@ export default class App extends Component {
       longitude: -100,
       zoom: 3,
       bearing: 0,
-      pitch: 0,
-      width: 500,
-      height: 500
+      pitch: 0
     }
   };
 
   componentDidMount() {
-    window.addEventListener('resize', this._resize);
-    this._resize();
     animation = window.requestAnimationFrame(this._animatePoint);
   }
 
   componentWillUnmount() {
-    window.removeEventListener('resize', this._resize);
     window.cancelAnimationFrame(animation);
   }
-
-  _resize = () => {
-    this.setState({
-      viewport: {
-        ...this.state.viewport,
-        width: this.props.width || window.innerWidth,
-        height: this.props.height || window.innerHeight
-      }
-    });
-  };
 
   _animatePoint = () => {
     this._updatePointData(pointOnCircle({center: [-100, 0], angle: Date.now() / 1000, radius: 20}));
@@ -77,6 +62,8 @@ export default class App extends Component {
     return (
       <MapGL
         {...viewport}
+        width="100%"
+        height="100%"
         mapStyle={mapStyle}
         onViewportChange={this._onViewportChange}
         mapboxApiAccessToken={MAPBOX_TOKEN} >
@@ -87,6 +74,6 @@ export default class App extends Component {
 
 }
 
-export function renderToDom() {
-  render(<App/>, document.body.appendChild(document.createElement('div')));
+export function renderToDom(container) {
+  render(<App/>, container);
 }