+++ website/src/components/examples.js
@@ -8,42 +8,12 @@ const HEIGHT_OFFSET = 64;
 
 export default class Examples extends Component {
 
-  constructor(props) {
-    super(props);
-    this.state = {
-      viewport: this._getSize()
-    };
-  }
-
-  componentDidMount() {
-    window.onresize = () => {
-      this.setState({
-        viewport: this._getSize()
-      });
-    };
-  }
-
-  componentWillUnmount() {
-    window.onresize = null;
-  }
-
-  _getSize() {
-    const {innerWidth, innerHeight} = window;
-
-    return {
-      width: innerWidth >= 576 ? innerWidth - WIDTH_OFFSET : innerWidth,
-      height: innerHeight - HEIGHT_OFFSET
-    };
-  }
-
   render() {
-    const {viewport} = this.state;
     const {route: {childComponent}} = this.props;
     const ExampleComponent = childComponent;
     return (
       <div className="flexbox-item flexbox-item--fill">
-        <ExampleComponent containerComponent={InfoPanel}
-          {...viewport} />
+        <ExampleComponent containerComponent={InfoPanel} />
       </div>
     );
   }